+++ common/repo/CMakeLists.txt
@@ -21,15 +21,16 @@
 # ------------   ----------   -----------------------------------------------
 # 2023-04-19     xqyjlj       initial version
 #
+set(name CspRepo)
 
 file(GLOB_RECURSE SOURCES
         inc/*.h
         src/*.cpp
         resources/*.qrc)
 
-add_library(repo STATIC ${SOURCES})
-target_link_libraries(repo PUBLIC compat_layer core CspConfig)
-target_include_directories(repo PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/inc)
+add_library(${name} STATIC ${SOURCES})
+target_link_libraries(${name} PUBLIC compat_layer core CspConfig)
+target_include_directories(${name} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/inc)
 
 ##########################################################################
 
@@ -40,6 +41,6 @@ foreach (TESTCASE ${TESTCASES})
     get_filename_component(TARGET_NAME ${TESTCASE} NAME_WE)
     add_executable(${TARGET_NAME} ${TESTCASE} test/res/res.qrc)
     target_compile_definitions(${TARGET_NAME} PRIVATE -DCSP_EXE_DIR=\"${CSP_EXE_DIR}\")
-    target_link_libraries(${TARGET_NAME} PRIVATE Qt5::Test repo)
+    target_link_libraries(${TARGET_NAME} PRIVATE Qt5::Test ${name})
     add_test(NAME ${TARGET_NAME} COMMAND ${TARGET_NAME})
 endforeach ()
